refactor: create customization to create polygon
